Fix typing/cast

+++ cubic-server/protocol/ClientPackets.cpp
@@ -79,8 +79,10 @@ std::shared_ptr<std::vector<uint8_t>> protocol::createSpawnPlayer(const SpawnPla
 std::shared_ptr<std::vector<uint8_t>> protocol::createEntityAnimation(EntityAnimationID animId, int32_t entityID)
 {
     std::vector<uint8_t> payload;
-    serialize(payload, entityID, addVarInt,
-        animId, addByte);
+    serialize(payload,
+        entityID, addVarInt,
+        (uint8_t)animId, addByte
+    );
     auto packet = std::make_shared<std::vector<uint8_t>>();
     finalize(*packet, payload, (int32_t) ClientPacketID::EntityAnimation);
     return packet;
